+++ src/main/java/kz/attractorschool/forum/domain/comment/CommentRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Integer> {
-  Page<Comment> findByTopic_id(Long topicId, Pageable pageable);
+  Page<Comment> findByTopic_id(Integer topicId, Pageable pageable);
 }
